Add isKeyboardShown command for iOS
