+++ src/main/java/io/appium/java_client/AppiumDriver.java
@@ -64,7 +64,7 @@ import java.util.Set;
 */
 @SuppressWarnings("unchecked")
 public class AppiumDriver<T extends WebElement>
-    extends DefaultGenericMobileDriver<T> implements ComparesImages {
+    extends DefaultGenericMobileDriver<T> implements ComparesImages, HasDeviceDetails {
 
     private static final ErrorHandler errorHandler = new ErrorHandler(new ErrorCodesMobile(), true);
     // frequently used command parameters
